Add initial data structures as well as very simple construction functions.
This commit adds a bunch of documentation and very little implementation for the
Span and SpannedStr structures. The difference between these two structures is
that Span does not hold the content itself. As such, SpannedStr *should* be used
during parsing while *Span* should be used to represent token position in the
generated AST.