Add the Lexer type
Generates a sequence of tokens from a SpannedStr, and holds them. It tries to recover when an error happens, so that as much errors as possible can be returned by Lexer::from_spanned_str.

Next commits will focus on the LexerState type, that will borrow a sequence of Tokens, perhaps from a Lexer, and clone them when necessary. It will represent a cursor on the input stream.