Make the token macro accept more meta items
Previous implementation of the token macro matcher allowed only two kind of meta items: documentation and derive macros. The downside of this method is that it does not allow for any additional item, and forces the user to specify the documentation first.

This commit removes such distinction. The meta items before the token name are applied both to the token product- and sum-type.